% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{addSearchTerm}
\alias{addSearchTerm}
\title{addSearchTerm}
\usage{
addSearchTerm(df, term)
}
\arguments{
\item{df}{the dataframe to add the term to}

\item{term}{a string containing the term to add}
}
\value{
the dataframe with a list added
}
\description{
Adds a search term to the search dataframe
}
\examples{
searchTerms <- addSearchTerm(searchTerms, "Paramedic AND Ambulance")
}
