% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanText}
\alias{cleanText}
\title{cleanText}
\usage{
cleanText(text)
}
\arguments{
\item{text}{a character vector containing the string we want to clean}
}
\value{
the cleaned text
}
\description{
cleans the abstract and/or full text to only show what we're interested in
}
\examples{
clean <- cleanText(dirty)
}
\keyword{internal}
