% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{importBibliography}
\alias{importBibliography}
\title{importBibliography}
\usage{
importBibliography(
  path,
  columns = c("author", "title", "abstract", "keywords", "Label"),
  keywords = TRUE
)
}
\arguments{
\item{path}{path to a bibliography file}

\item{columns}{the columns to extract}

\item{keywords}{whether or not to add keyword column}
}
\value{
the bibliography data frame
}
\description{
imports a bibliography from a file
uses synthesisr under the hood, so accepts
any filetypes synthesisr will accept
}
\examples{
bib <- importBibliography("./inst/extdata/test.ris")
}
