% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{initialiseSearchTerms}
\alias{initialiseSearchTerms}
\title{initialiseSearchTerms}
\usage{
initialiseSearchTerms()
}
\value{
dataframe containing an empty history column
}
\description{
Initialise the search terms dataframe
}
\examples{
st <- initialiseSearchTerms()
}
