% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{listUniqueWordsInBibliography}
\alias{listUniqueWordsInBibliography}
\title{listUniqueWordsInBibliography}
\usage{
listUniqueWordsInBibliography(
  bibframe,
  columns = c("title", "abstract", "keywords")
)
}
\arguments{
\item{bibframe}{the bibliography dataframe}

\item{columns}{the columns to include}
}
\value{
a character vector containing the unique words in the bibliography
}
\description{
creates a character vector containing the unique words in a bibliography
}
\examples{
unique1 <- listUniqueWordsInBibliography(bibframe)
}
